requirements.txt
asgiref==3.2.3
astroid==2.3.3
cffi==1.13.2
cryptography==2.8
dj-database-url==0.5.0
Django==3.0.1
gunicorn==20.0.4
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2==2.8.4
pycparser==2.19
pylint==2.4.4
pyOpenSSL==19.1.0
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
whitenoise==5.0.1
wrapt==1.11.2
